requirements.txt
deep-translator
scholarly
transformers
torch
python-dotenv
